Hardcoded in Facebook App ID and Secret Key*****
Facebook login and signup should now work. I have not written any tests
for this yet.
*****This is not secure. People can see the secret key in our github and
use it to authorize requests "from our website". In the future, I'll
figure out how to create a global variable or service or something that
reads the key from a secure text file and can be referenced in our code
without showing the actual key.