Calendar testing now supports data validation
The tests are viewable with `grunt protractor:calendar`

I don't really know what data validation means but it seems legit. The
calendar tests can now check the calendar entries to verify that the
title/date/description is correct.

If you actually look through my tests, you might wonder why we expect to
see the test code as "whatever\nDelete". This is because someone
(*cough* Max) included the Delete button under the same paragraph tag as
the description text so poor little me was unable to separate the two.
So, since the delete button is always going to be there, I just added it
to the expectation.