Optimized login specs and page objects
Just cleaning up the code and syntax.

Eliminated use of vars in the page objects -- from my personal
experience, it's WAY better to use functions. I ran into a huge error
where my page object was trying to grab elements on the page but the
page wasn't in the browser yet. Using functions instead of variables in
the page object prevents this error.