Updated noAuth & user page object
Had to add to the page objects for noAuth and user for interaction with
the purchase modal. I will be adding tests for interaction with these
components when I get a chance.

Updated one test in noAuth to expect a title of "Karate 1" instead of
"Karate".

I didn't add purchase modal interaction to the instructor and admin page
objects because these roles won't need to purchase courses.