Protractor:courses fixes
There is a bug with one of the tests testing to see if the purchase
button disappears. I will fix this later. It's because the button isn't
ng hidden, the div that it's inside of is hidden (so I have to test to
see if the div is hidden instead of the button)